+++ src/utils/index.ts
@@ -1,11 +1,16 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import * as parser from '@babel/parser';
+import { FileFilter, PathMatcher } from '../types';
 
 export const readFile = (i: string): string => fs.readFileSync(i, 'utf8');
 
-export const listFiles = (directory: string, visited: { [prop: string]: true } = {}): string[] => {
-    if (visited[directory]) {
+export const innerListFiles = (
+    directory: string,
+    exclude: PathMatcher,
+    visited: { [prop: string]: true },
+): string[] => {
+    if (visited[directory] || exclude(directory)) {
         return [];
     }
 
@@ -19,7 +24,7 @@ export const listFiles = (directory: string, visited: { [prop: string]: true } =
 
     // TODO: add support for multiple symlinks pointing to the same location
     // Currently, if wee have already seen the location, we will not look into it (both for symlinks and normal directories)
-    if (visited[realPath]) {
+    if (visited[realPath] || exclude(realPath)) {
         return [];
     }
 
@@ -33,9 +38,36 @@ export const listFiles = (directory: string, visited: { [prop: string]: true } =
 
     const files = fs
         .readdirSync(directory)
-        .reduce((acc, f) => [...acc, ...listFiles(path.join(directory, f), visited)], []);
+        .reduce(
+            (acc, f) => [...acc, ...innerListFiles(path.join(directory, f), exclude, visited)],
+            [],
+        );
 
     return files;
+};
+
+export const getMatcher = (filter: RegExp | RegExp[] | Function): PathMatcher => {
+    if (typeof filter === 'function') {
+        return path => filter(path);
+    }
+
+    const regexSetting = Array.isArray(filter) ? filter : [filter];
+
+    return path => regexSetting.some(i => i.test(path));
+};
+
+export const listFiles = (directory: string, exclude?: FileFilter): string[] => {
+    let excludeMatcher: PathMatcher = () => false;
+
+    if (exclude) {
+        const matcher = getMatcher(exclude);
+        const rootToReplace = directory + path.sep;
+        excludeMatcher = filePath => matcher(filePath.replace(rootToReplace, ''));
+    }
+
+    const result = innerListFiles(directory, excludeMatcher, {});
+
+    return result;
 };
 
 const defaultPlugins = [